Unify BMPCoords
Two classes deleted! One class! Two class!
Ha. Ha. Ha.
